# Reduce single file format overhead by lazily importing modules and loading .gitignore / --exclude