@ichard26 2022-08-05
For what it is worth, I'd prefer a rebase merge for this PR given the commits are mostly independent (other than the changelog entry).