@ichard26 2022-08-27
This works on Windows with PyInstaller (just tested locally) so I'm merging.